rename packageethclient_utils as ethclient + resolve package name conflicts with subnet-evm/ethclient package
