+++ relayer/application_relayer.go
@@ -23,7 +23,7 @@ import (
 	avalancheWarp "github.com/ava-labs/avalanchego/vms/platformvm/warp"
 	"github.com/ava-labs/awm-relayer/config"
 	"github.com/ava-labs/awm-relayer/database"
-	"github.com/ava-labs/awm-relayer/ethclient_utils"
+	"github.com/ava-labs/awm-relayer/ethclient"
 	"github.com/ava-labs/awm-relayer/messages"
 	"github.com/ava-labs/awm-relayer/peers"
 	"github.com/ava-labs/awm-relayer/utils"
@@ -639,7 +639,7 @@ func (r *applicationRelayer) calculateStartingBlockHeight(processHistoricalBlock
 
 // Gets the height of the chain head, writes it to the database, then returns it.
 func (r *applicationRelayer) setProcessedBlockHeightToLatest() (uint64, error) {
-	ethClient, err := ethclient_utils.DialWithConfig(
+	ethClient, err := ethclient.DialWithConfig(
 		context.Background(),
 		r.sourceBlockchain.RPCEndpoint,
 		r.sourceBlockchain.HttpHeaders,
